requirements.txt
django
gunicorn
pymysql
mysqlclient
#mysql-connector-python
reportlab
django-crum
django-model-utils
pillow<=5.0.0,>=2.4.0
django-versatileimagefield
python-decouple
xhtml2pdf>=0.2b1
django-easy-pdf
WeasyPrint
django-debug-toolbar
django-developer-panel
django-extensions
django-autocomplete-light
django-images
cmdln
dj-inmemorystorage
djangorestframework
requests
six
#flask
#redis
